+++ tests/test_visualize.py
@@ -1,0 +1,12 @@
+import unittest
+import doctest
+import moxel.visualize
+
+
+def load_tests(loader, tests, ignore):
+    tests.addTests(doctest.DocTestSuite(moxel.visualize))
+    return tests
+
+
+if __name__ == '__main__':
+    unittest.main()